+++ previous-versions/common/src/sdk/models/operations/configuresupplementaldata.ts
@@ -76,5 +76,5 @@ export class ConfigureSupplementalDataResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }