+++ previous-versions/common/src/sdk/models/operations/getprofile.ts
@@ -35,5 +35,5 @@ export class GetProfileResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }