+++ previous-versions/common/src/sdk/models/operations/getprofilesyncsettings.ts
@@ -29,7 +29,7 @@ export class GetProfileSyncSettingsResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * OK