+++ previous-versions/common/src/sdk/models/operations/getpushoperation.ts
@@ -49,5 +49,5 @@ export class GetPushOperationResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }