+++ previous-versions/common/src/sdk/sdk.ts
@@ -61,9 +61,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "3.0.0";
-    sdkVersion = "0.40.0";
-    genVersion = "2.188.1";
-    userAgent = "speakeasy-sdk/typescript 0.40.0 2.188.1 3.0.0 @codat/common";
+    sdkVersion = "0.40.1";
+    genVersion = "2.194.1";
+    userAgent = "speakeasy-sdk/typescript 0.40.1 2.194.1 3.0.0 @codat/common";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);