+++ previous-versions/common/RELEASES.md
@@ -701,3 +701,13 @@ Based on:
 - [typescript v0.40.0] previous-versions/common
 ### Releases
 - [NPM v0.40.0] https://www.npmjs.com/package/@codat/common/v/0.40.0 - previous-versions/common
+
+## 2023-11-20 01:30:43
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.121.1 (2.194.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.40.1] previous-versions/common
+### Releases
+- [NPM v0.40.1] https://www.npmjs.com/package/@codat/common/v/0.40.1 - previous-versions/common